refactor: style 관련 코드를 분리한다.
